Enable warning 4389 as an error on windows builds.  This will make windows builds more similar to linux/mac, which already treat signed/ unsigned equality comparisons as warnings (and hence errors).
BUG=44471
TEST=none

Review URL: http://codereview.chromium.org/2081007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@48186 0039d316-1c4b-4281-b951-d872f2087c98

+++ app/os_exchange_data_provider_win.cc
@@ -1,5 +1,5 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2009 The Chromium Authors. All rights reserved.
+// Copyright (c) 2010 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -85,7 +85,7 @@ class FormatEtcEnumerator : public IEnumFORMATETC {
   std::vector<FORMATETC*> contents_;
 
   // The cursor of the active enumeration - an index into |contents_|.
-  int cursor_;
+  size_t cursor_;
 
   LONG ref_count_;
 
@@ -126,9 +126,8 @@ STDMETHODIMP FormatEtcEnumerator::Next(
     DCHECK(count == 1);
 
   // This method copies count elements into |elements_array|.
-  int index = 0;
-  while (cursor_ < static_cast<int>(contents_.size()) &&
-         static_cast<ULONG>(index) < count) {
+  ULONG index = 0;
+  while (cursor_ < contents_.size() && index < count) {
     CloneFormatEtc(contents_.at(cursor_), &elements_array[index]);
     ++cursor_;
     ++index;
@@ -145,7 +144,7 @@ STDMETHODIMP FormatEtcEnumerator::Skip(ULONG skip_count) {
   cursor_ += skip_count;
   // MSDN implies it's OK to leave the enumerator trashed.
   // "Whatever you say, boss"
-  return cursor_ <= static_cast<int>(contents_.size()) ? S_OK : S_FALSE;
+  return cursor_ <= contents_.size() ? S_OK : S_FALSE;
 }
 
 STDMETHODIMP FormatEtcEnumerator::Reset() {
